Adding documentation to data_kit
Documentation for models submodule is still missing. Next step for me.
