+++ modules/DMZ/data_kit/data_splitting.py
@@ -1,6 +1,6 @@
 import pandas as pd
-from sklearn import preprocessing
 from sklearn import cross_validation
+import data_prepping
 import numpy as np
 
 
@@ -17,7 +17,7 @@ def get_train_test(pandas_data, target_col):
 
     # Separating target from the rest of the data
     x = pandas_data.drop(target_col, 1)
-    x = scale_numeric_data(x)
+    x = data_prepping.scale_numeric_data(x)
 
     # Selection of training/target data for validation and training.
     target_loc = pandas_data.columns.get_loc(target_col)
@@ -40,20 +40,3 @@ def separate_target(pandas_data, target_col):
     x = pd.DataFrame.as_matrix(x)
 
     return x, y
-
-
-# Scales the numeric input data.
-# Input:
-# - Pandas Dataframe, non-scaled
-# Output:
-# - Pandas Dataframe that has been scaled.
-def scale_numeric_data(pandas_data):
-    # Scaling is important because if the variables are too different from
-    # one another, it can throw off the model.
-    # EX: If one variable has an average of 1000, and another has an average
-    # of .5, then the model won't be as accurate.
-    for col in pandas_data.columns:
-        if pandas_data[col].dtype == np.float64 or pandas_data[col].dtype == np.int64:
-            pandas_data[col] = preprocessing.scale(pandas_data[col])
-
-    return pandas_data
